requirements.txt
langchain==0.1.17
langchain-community==0.0.24
openai
faiss-cpu
pdfplumber
python-docx
nltk
tqdm
python-dotenv